Moved the loading screen into gallery page and made page layout much better for styling purposes
