+++ Desktop/src/components/Spinner.svelte
@@ -29,8 +29,8 @@
     justify-content: center;
     align-items: center;
     height: 100vh;
-    width: 100vw;
-    background-color: #f0f0f0;
+    width: 100vh;
+    background-color: inherit;
   }
 
   .container {
