+++ Desktop/src/routes/ProtectedRoutes.svelte
@@ -5,11 +5,6 @@
   import { navigate } from "@sveltejs/kit/navigation"; // If using SvelteKit for navigation
   import Sidebar from "../components/Sidebar.svelte";
 
-  import { isLoading } from "../stores/loading";
-  import Spinner from "../components/Spinner.svelte";
-
-  let data = null;
-
   let authToken;
 
   // Subscribe to the token store to get the current authentication state
@@ -24,38 +19,15 @@
       navigate("#login"); // Change this to your login route
     }
   });
-
-  onMount(async () => {
-    isLoading.set(true);
-    try {
-      // Simulate data fetching with a delay
-      await new Promise((resolve) => setTimeout(resolve, 3000));
-      data = await fetchData();
-    } catch (error) {
-      console.error("Failed to fetch data", error);
-    } finally {
-      isLoading.set(false);
-    }
-  });
-
-  async function fetchData() {
-    // Replace with your actual data fetching logic
-    return { message: "Data loaded successfully" };
-  }
 </script>
 
 <!-- Slot to render the protected content if authenticated -->
 
-<div class="flex">
-  <div class="w-1/5">
+<div class="grid grid-cols-5">
+  <div class="col-span-1">
     <Sidebar />
   </div>
-  {#if $isLoading}
-    <Spinner />
-  {/if}
-  {#if data}
-    <div class="flex-1 items-center p-4">
+    <div class="col-span-4">
       <slot />
     </div>
-  {/if}
 </div>
